+++ src/ndf.cpp
@@ -36,7 +36,7 @@ void ndf::_bind_methods() {
 numdot::reduction_new<double_t>([](const va::VArrayTarget& target, const va::VArray& array) { va::func(va::store::default_allocator, target, array.data, nullptr); }, (varray1))
 
 #define REDUCTION2(func, varray1, varray2) \
-numdot::reduction_new<double_t>([](const va::VArrayTarget& target, const va::VArray& x1, const va::VArray& x2) { va::func(va::store::default_allocator, target, x1.data, x2.data, nullptr); }, (varray1), (varray2))
+numdot::reduction_new_binary_weak<double_t>([](const va::VArrayTarget& target, const va::VArray& x1, const va::VArray& x2) { va::func(va::store::default_allocator, target, x1.data, x2.data, nullptr); }, (varray1), (varray2))
 
 double_t ndf::sum(const Variant& a) {
 	return REDUCTION1(sum, a);